refactor(drupal/cypress): simplify step definitions, auto-uninstall drupal
