+++ packages/composer/drupal/cypress/tests/Cypress/integration/Session/index.js
@@ -1,6 +1,35 @@
 // @ts-check
 /// <reference types="Cypress" />
-import {langCodes} from "../steps";
+
+const langCodes = {
+  'German': 'de'
+};
+
+beforeEach(() => {
+  cy.drupalInstall({
+    setup: 'cypress:integration/CypressTestSiteInstallScript.php',
+  });
+});
+
+// Given a test case uses 'cy.drupalSession' to authenticate in as "admin"
+Given(/^the test case uses 'cy.drupalSession' to authenticate as "([^"]*)"$/, function (account) {
+  cy.drupalSession({user: account});
+});
+
+// And the test case uses 'cy.drupalSession' to display switch to "German"
+And(/^the test case uses 'cy.drupalSession' to display switch to "([^"]*)"$/, function (language) {
+  cy.drupalSession({language: langCodes[language]});
+});
+
+// And the test case uses 'cy.drupalSession' to switch to workspace "stage"
+And(/^the test case uses 'cy.drupalSession' to switch to workspace "([^"]*)"$/, function (workspace) {
+  cy.drupalSession({workspace: workspace});
+});
+
+// And the test case uses 'cy.drupalSession' to display the toolbar
+Given(/^the test case uses 'cy.drupalSession' to display the toolbar$/, function () {
+  cy.drupalSession({toolbar: true});
+});
 
 // When the test case visits the homepage and clicks the link to the "admin" account
 When(/^the test case visits the homepage and clicks the link to the "([^"]*)" account$/, function (account) {
